Fix potential overflow in netstandard2.0 ReadString bounds check
Use (ulong) casts to match the overflow-safe pattern already used
by GetSpan on non-netstandard2.0 targets.

+++ MaxMind.Db/MemoryMapBuffer.cs
@@ -353,7 +353,7 @@ namespace MaxMind.Db
             {
                 return string.Empty;
             }
-            if (offset < 0 || offset + count > Length)
+            if (offset < 0 || (ulong)offset + (ulong)count > (ulong)Length)
             {
                 throw new ArgumentOutOfRangeException(nameof(offset),
                     "Attempt to read beyond the end of the MemoryMappedFile.");
